Todo removed from project; Rotation of pixmap now being tracked by a float; SLOTS and SIGNALS connected for clicking an event

+++ monthlabel.cpp
@@ -39,7 +39,7 @@ QPixmap MonthLabel::paintMonthsOnPixmap(QPixmap pixmap)
 
 void MonthLabel::slotGrabMouseMove()
 {
-    int cRotation = getCurrentRotation();
+    float cRotation = getCurrentRotation();
     if(cRotation < 0)
         cRotation += 360;
     qDebug() << cRotation ;
